% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Symlink the curatedMetagenomicData Command-Line Interface}
\usage{
install(linked_file = "/usr/local/bin/curatedMetagenomicData")
}
\arguments{
\item{linked_file}{the path to the linked file that will be created,
\code{"/usr/local/bin/curatedMetagenomicData"} by default}
}
\value{
the exit status of the system command, invisibly
}
\description{
The command-line interface is an executable R script that lives in the
curatedMetagenomicDataTerminal package. It must be symlinked to a location on
your \verb{$PATH} to be used from a terminal - the \code{install()} function does this
by making use of the \code{ln} system command. So, this won't work if you have
Windows, but get in touch if you need it to.
}
\examples{
\dontrun{
install()
}
}
